# server/docs/swagger.yaml
definitions:
  apis.ReqLogin:
    properties:
      password:
        description: 密码
        type: string
      phone:
        description: 手机号
        type: string
    type: object
  apis.ReqOrders:
    properties:
      direction:
        description: 类型 0买 1卖
        type: integer
      price:
        description: 价格(分)
        type: integer
      volume:
        description: 几手
        type: integer
    type: object
  apis.ReqRegister:
    properties:
      code:
        description: 验证码
        type: string
      password:
        description: 密码
        type: string
      phone:
        description: 手机号
        type: string
    type: object
  apis.ReqSms:
    properties:
      phone:
        description: 手机号
        type: string
    type: object
  apis.ReqUpdatePassword:
    properties:
      old_password:
        description: 旧密码
        type: string
      password:
        description: 密码
        type: string
    type: object
  apis.ReqUpdatePhone:
    properties:
      code:
        description: 验证码
        type: string
      phone:
        description: 手机号
        type: string
    type: object
  apis.ReqUpdateUser:
    properties:
      avatar:
        description: 头像
        type: string
      nickname:
        description: 昵称
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: 接口文档
  title: 中鼎接口API
  version: 0.0.1
paths:
...
  /api/file/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 上传文件示例
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 上传文件示例,返回包括文件详情
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 上传文件
      tags:
      - 前端接口API
...
  /api/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登录
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/apis.ReqLogin'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"data":{"access_token":"ZWdUJTJCUGliYWNZQ0NGVUwzQUtUbFZwd3ZlR3FDWURGMFFlZUdMWFJaQzdMY3BWbGxiUnR6VjJOc0JYVmRBMEpsdlJCayUyQjVLWHUxNHRLVXUwekZ5dE5nJTNEJTNE"},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 登录
      tags:
      - 前端接口API
  /api/orders/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: 下单参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/apis.ReqOrders'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"data":{},"msg":"success"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: 下单
      tags:
      - 前端接口API
  /api/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 注册
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/apis.ReqRegister'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"创建成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 注册
      tags:
      - 前端接口API
  /api/sms/send:
    post:
      consumes:
      - application/json
      parameters:
      - description: 获取验证码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/apis.ReqSms'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"创建成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 获取短信验证码
      tags:
      - 前端接口API
...
  /api/user/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: 更新用户信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/apis.ReqUpdateUser'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"data":{},"msg":"success"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: 更新用户信息
      tags:
      - 前端接口API
  /api/user/update-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: 更新密码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/apis.ReqUpdatePassword'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"data":{},"msg":"success"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: 更换密码
      tags:
      - 前端接口API
  /api/user/update-phone:
    post:
      consumes:
      - application/json
      parameters:
      - description: 更新手机号
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/apis.ReqUpdatePhone'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"data":{},"msg":"success"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: 更换手机号
      tags:
      - 前端接口API
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"